refactor text-concave-cutout test
